requirements.txt
Django~=3.1.1
djangorestframework~=3.11.1
psycopg2-binary==2.8.4
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.5.0
